% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.r
\name{bootstrap}
\alias{bootstrap}
\alias{bootstrap,mfa-method}
\title{Bootstrap}
\usage{
bootstrap(object, nbt = 1000)
}
\arguments{
\item{object}{an 'mfa' object}

\item{nbt}{an integer number of a bootstrap sample size}
}
\value{
a matrix of bootstrap ratio
}
\description{
Return a matrix of bootstrap ratio
}
\examples{
# default 
ndatas<-apply(wine_data,2,function(x){ (x-mean(x))/norm(x-mean(x),type="2")})
test<-mfa(ndatas,sets=list(1:6,7:12,13:18,19:23,24:29,30:34,35:38,39:44,45:49,50:53),center=FALSE,scale=FALSE)
bootstrap(test,1000)
}

