% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap.mfa}
\alias{bootstrap.mfa}
\title{bootstrap.mfa}
\usage{
bootstrap.mfa(X, L = 10)
}
\arguments{
\item{X}{a mfa object}

\item{L}{number of bootstrap samples are computed}
}
\value{
bootstrap estimates and bootstrap ratio
}
\description{
main function of bootstrap
}
\examples{
\dontrun{bootstrap.mfa(mfa.obj, 1000)}
}

