% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{bootstrap}
\usage{
bootstrap(mfa, value)
}
\arguments{
\item{mfa}{the mfa object created by the mfa function}

\item{value}{the number of times of bootstrap}
}
\value{
a list, containing the mean, standard deviation and bootstrap ratio; also, the plot of bootstrap ratio for each dimension is created
}
\description{
method of bootstrap for mfa object
}
\examples{

wines <- read.csv("https://raw.githubusercontent.com/ucb-stat243/stat243-fall-2016/master/problem-sets/final-project/data/wines.csv", stringsAsFactors = FALSE)
sets <- list(2:7, 8:13, 14:19, 20:24, 25:30, 31:35, 36:39, 40:45, 46:50, 51:54)
scaling_vec <- apply(subset(wines, select = unlist(sets)), 2, function(x) sqrt(sum((x - mean(x))^2)))
mymfa1 <- mfa(wines, sets, ncomps = 2, T, scaling_vec)

bootstrap(mymfa1,1000)
}

