% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{bootstrap_factorscore}
\alias{bootstrap_factorscore}
\title{Bootstrapping Factor Scores}
\usage{
bootstrap_factorscore(x)
}
\arguments{
\item{x}{An object of class mfa}
}
\value{
b bootstrap factorscore
from the set of tables. This approach also computes boostrap ratios for
each dimension.
}
\description{
Calculates the bootstrap confidence intervals by sampling
}
\details{
details
}

