% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_factorscores.R
\name{bootstrap_factorscores}
\alias{bootstrap_factorscores}
\title{Bootstrapping Factor Scores}
\usage{
bootstrap_factorscores(x)
}
\arguments{
\item{x}{An object of class mfa}
}
\description{
Calculates the bootstrap confidence intervals by sampling
}
\examples{
b bootstrap factorscore
from the set of tables. This approach also computes boostrap ratios for
each dimensionnnnnnnnnnn.

}

