% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{BootstrapRatios}
\alias{BootstrapRatios}
\title{Bootstrap Ratios for Factor Scores}
\usage{
BootstrapRatios(mfa, times = 1000)
}
\arguments{
\item{mfa}{a "mfa" object returned by function mfa}

\item{times}{the size of bootstrap sample}
}
\value{
a matrix of bootstrap ratios for factor scores
}

