% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{calculate_chi}
\alias{calculate_chi}
\title{Calculate posterior precision parameters}
\usage{
calculate_chi(m)
}
\arguments{
\item{m}{A fit returned from \code{mfa}}
}
\description{
Calculate posterior precision parameters
}

