% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchByEffort.R
\name{catchByEffort}
\alias{catchByEffort}
\title{catchByEffort}
\usage{
catchByEffort(con, effortData, Special_Legal_Sizes = NULL)
}
\arguments{
\item{con}{Database connection.}

\item{effortData}{Survey effort data in SQL format from \code{\link[=FISH_query]{FISH_query()}}.}

\item{Special_Legal_Sizes}{Optional named vector of species-specific
legal sizes that overrides the default legal sizes.}
}
\value{
A data frame containing catch information summarized by survey,
effort, and species.
}
\description{
Internal function to calculate catch by effort
}
\details{
This function is used internally by \code{\link[=FISH_query]{FISH_query()}} to retrieve and
summarize catch data associated with survey effort. It was split out
from \code{\link[=FISH_query]{FISH_query()}} because of the complicated join across the three
catch tables.
}
\keyword{internal}
