% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchSummary.R
\name{catchSummary}
\alias{catchSummary}
\title{catchSummary}
\usage{
catchSummary(catchData)
}
\arguments{
\item{catchData}{A data frame returned by \code{\link[=FISH_query]{FISH_query()}} with
\code{QueryType = "Catch"}. The data must contain \code{SurveyId}, \code{Species},
\code{GearType}, \code{TotalNumberCaught}, \code{LengthAverage}, \code{LengthMinimum},
\code{LengthMaximum}, and \code{N_legal}.}
}
\value{
A data frame with one row per survey and species combination.
Columns include total catch, weighted average length, minimum and
maximum length, number of legal fish, and percentage of fish that
meet  statewide or other specified legal minimum size limit.
}
\description{
Summarizes catch data returned by \code{\link[=FISH_query]{FISH_query()}} to produce one row
for each survey and species combination. The summary includes total
catch, weighted average length, minimum and maximum length, and the
number and percentage of fish meeting the statewide or other specified
legal minimum size limit.
}
\examples{
\dontrun{
#connect to database
con <- FISHub_connect()

catchData<-FISH_query(con,QueryType = "Catch",SurveyId = 805)
catchSum <- catchSummary(catchData)

#' #disconnect from database
DBI::dbDisconnect(con)
}

}
