% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_B.R
\name{check_B}
\alias{check_B}
\title{check_B}
\usage{
check_B(B)
}
\arguments{
\item{B}{the (hopefully) numeric integer to test}
}
\value{
boolean - does the test pass?
}
\description{
check that the B input is valid
}

