% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_center.R
\name{check_center}
\alias{check_center}
\title{check_center}
\usage{
check_center(center)
}
\arguments{
\item{center}{boolean or numeric vector}
}
\value{
boolean - does the check pass?
}
\description{
make sure center has a viable value
}

