% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_contribution_params.R
\name{check_contribution_params}
\alias{check_contribution_params}
\title{check_contribution_params}
\usage{
check_contribution_params(mfa, l_max)
}
\arguments{
\item{mfa}{- a valid mfa object}

\item{l_max}{- numeric integer}
}
\description{
make sure the inputs to contribution_table_dim are
viable
}

