% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{check_data}
\usage{
check_data(data, sets, ncomps, weights, ids)
}
\arguments{
\item{data}{- the data to check}

\item{sets}{- the blocks the data should be broken into}

\item{ncomps}{- integer number of components to use}

\item{weights}{- numeric vector of weights}

\item{ids}{- character/numeric vector of sample id's}
}
\value{
boolean - TRUE if test passes
}
\description{
private function to make sure data is entered correctly
}

