% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ids.R
\name{check_ids}
\alias{check_ids}
\title{check_ids}
\usage{
check_ids(ids)
}
\arguments{
\item{ids}{- character vector}
}
\value{
boolean - did the test pass?
}
\description{
check that the id variable is viable
}

