% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ncomps.R
\name{check_ncomps}
\alias{check_ncomps}
\title{check_ncomps}
\usage{
check_ncomps(ncomps)
}
\arguments{
\item{ncomps}{- numeric integer value number of component}
}
\value{
boolean - did the test pass
}
\description{
make sure ncomps has a viable value
}

