% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plot_dims.R
\name{check_plot_dims}
\alias{check_plot_dims}
\title{check_plot_dims}
\usage{
check_plot_dims(nGroups, dims)
}
\arguments{
\item{nGroups}{- numberic value of how many total groups there are}

\item{dims}{- numeric vector of dimensions}
}
\value{
boolean - did the test pass
}
\description{
make sure plot dimensions has a viable value
}

