% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_scale.R
\name{check_scale}
\alias{check_scale}
\title{check_scale}
\usage{
check_scale(scale)
}
\arguments{
\item{scale}{boolean or numeric vector}
}
\value{
boolean - does the check pass?
}
\description{
make sure scale has a viable value
}

