% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_seed.R
\name{check_seed}
\alias{check_seed}
\title{check_seed}
\usage{
check_seed(seed)
}
\arguments{
\item{seed}{the (hopefully) numeric integer to test}
}
\value{
boolean - does the test pass?
}
\description{
check that the seed input is valid
}

