% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sets.R
\name{check_sets}
\alias{check_sets}
\title{check_sets}
\usage{
check_sets(sets)
}
\arguments{
\item{sets}{- list of numeric/character vectors}
}
\value{
boolean - did the test pass?
}
\description{
make sure sets has viable value
}

