% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_weights.R
\name{check_weights}
\alias{check_weights}
\title{check_weights}
\usage{
check_weights(weights)
}
\arguments{
\item{weights}{- numeric vector or null}
}
\value{
boolean - does the test pass?
}
\description{
check that the weights variable is viable
}

