% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{compromise_plot}
\alias{compromise_plot}
\alias{compromise_plot,mfa-method}
\title{compromise_plot}
\usage{
compromise_plot(x)

\S4method{compromise_plot}{mfa}(x)
}
\arguments{
\item{x}{a `mfa` object}
}
\value{
A plot of the compromises
}
\description{
compromise_plot
}
\section{Methods (by class)}{
\itemize{
\item \code{mfa}: plot of the tables
}}

