% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{contri_obs}
\alias{contri_obs}
\alias{contri_obs.mfa}
\title{Contribution of observations}
\usage{
contri_obs(x)

\method{contri_obs}{mfa}(x)
}
\arguments{
\item{x}{mfa object}
}
\value{
a matrix with its [i,j] element be the contribution of the ith observation to the jth component
}
\description{
Calculates the contributions of the observations to the extracted components (dimensions)
}
\examples{
contri_obs(wine)
}

