% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{contri_table}
\alias{contri_table}
\alias{contri_table.mfa}
\title{Contribution of subtables}
\usage{
contri_table(x)

\method{contri_table}{mfa}(x)
}
\arguments{
\item{x}{mfa object}
}
\value{
a matrix with its [i,j] element be the contribution of the ith subtable to the jth component
}
\description{
Calculates the contributions of the subtables to the extracted components (dimensions)
}
\examples{
contri_table(wine)
}

