% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{contri_var}
\alias{contri_var}
\alias{contri_var.mfa}
\title{Contribution of variables}
\usage{
contri_var(x)

\method{contri_var}{mfa}(x)
}
\arguments{
\item{x}{mfa object}
}
\value{
a matrix with its [i,j] element be the contribution of the ith variable to the jth component
}
\description{
Calculates the contributions of the variables to the extracted components (dimensions)
}
\examples{
contri_var(wine)
}

