% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{contrib_obs}
\alias{contrib_obs}
\title{contrib_obs}
\usage{
contrib_obs(mfaex)
}
\arguments{
\item{mfaex, }{an object of class "mfa"}
}
\value{
the matrix of contributions M[i,l] is the contribution of observation i to dimension l
}
\description{
Computes the contribution of each observation to each dimension \code{"contrib_obs"}
}
\examples{

wines <- read.csv("https://raw.githubusercontent.com/ucb-stat243/stat243-fall-2016/master/problem-sets/final-project/data/wines.csv", stringsAsFactors = FALSE)
sets <- list(2:7, 8:13, 14:19, 20:24, 25:30, 31:35, 36:39, 40:45, 46:50, 51:54)
scaling_vec <- apply(subset(wines, select = unlist(sets)), 2, function(x) sqrt(sum((x - mean(x))^2)))
mymfa1 <- mfa(wines, sets, ncomps = 2, center = TRUE, scale = scaling_vec)

contrib_obs1 <- contrib_obs(mymfa1)
}

