% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{contrib_tables}
\alias{contrib_tables}
\title{contrib_tables}
\usage{
contrib_tables(mfaex, setex, tables)
}
\arguments{
\item{mfaex, }{an object of class "mfa" builded from an original data-set containing the 'tables'}

\item{setex, }{the sets of variables used in the MFA (mapping between the tables and the mfa variables)}

\item{tables, }{some original tables}
}
\value{
the matrix of contributions M[k,l] is the contribution of table k to dimension l
}
\description{
Computes the contribution of each table to each dimension \code{"contrib_tables"}
}
\examples{

wines <- read.csv("https://raw.githubusercontent.com/ucb-stat243/stat243-fall-2016/master/problem-sets/final-project/data/wines.csv", stringsAsFactors = FALSE)
sets <- list(2:7, 8:13, 14:19, 20:24, 25:30, 31:35, 36:39, 40:45, 46:50, 51:54)
scaling_vec <- apply(subset(wines, select = unlist(sets)), 2, function(x) sqrt(sum((x - mean(x))^2)))
tables <- data_tables(wines, sets, TRUE, scaling_vec)
mymfa1 <- mfa(wines, sets, ncomps = 2, center = TRUE, scale = scaling_vec)

sets <- lapply(sets, FUN=function(vec){vec-1})
contrib_table1 <- contrib_tables(mymfa1, sets, tables)
}

