% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contributions.R
\name{ContribObs}
\alias{ContribObs}
\title{The contributions of an Observation to a Dimension for a given mfa object.}
\usage{
ContribObs(mfa)
}
\arguments{
\item{mfa}{Object of class 'mfa', which should be output from mfa function.}
}
\value{
A matrix of the contributions of each Observation to each
    Dimension, with rows corresponding to Observations, and columns
    corresponding to Dimensions.
}
\description{
The contributions of an Observation to a Dimension for a given mfa object.
}

