% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contributions.R
\name{ContribTable}
\alias{ContribTable}
\title{The contributions of a Table to a Dimension for a given mfa object.}
\usage{
ContribTable(mfa)
}
\arguments{
\item{mfa}{Object of class 'mfa', which should be output from mfa function.}
}
\value{
A matrix of the contributions of each Table to each
    Dimension, with rows corresponding to Tables, and columns
    corresponding to Dimensions.
}
\description{
The contributions of a Table to a Dimension for a given mfa object.
}

