% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution_obs_dim.R
\name{contribution_obs_dim}
\alias{contribution_obs_dim}
\title{contribution_obs_dim}
\usage{
contribution_obs_dim(mfa, l_max = 2)
}
\arguments{
\item{mfa}{- the mfa object}

\item{l_max}{- how many dimensions should be returned?}
}
\value{
a matrix with the contributions of an observation to a dimension
}
\description{
Calculate the contributions of an observation to a dimension
}

