% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution_table_dim.R
\name{contribution_table_dim}
\alias{contribution_table_dim}
\title{contribution_table_dim}
\usage{
contribution_table_dim(mfa, ...)
}
\arguments{
\item{mfa}{- the mfa object}

\item{l_max}{- how many dimensions should be returned?}
}
\value{
a matrix with the contributions of a table to a dimension
}
\description{
Calculate the contributions of a table to a dimension
}

