% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution_var_dim.R
\name{contribution_var_dim}
\alias{contribution_var_dim}
\title{contribution_var_dim}
\usage{
contribution_var_dim(mfa, l_max = 2)
}
\arguments{
\item{mfa}{- the mfa object}

\item{l_max}{- how many dimensions should be returned?}
}
\value{
a matrix with the contributions of a variable to a dimension
}
\description{
Calculate the contributions of a variable to a dimension
All contributions are scaled by multiplying by 1000
}

