% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.r
\docType{methods}
\name{contributions}
\alias{contributions}
\alias{contributions,}
\alias{contributions,mfa-method}
\alias{mfa-method}
\title{Contributions method}
\usage{
contributions(object)

\S4method{contributions}{mfa}(object)
}
\arguments{
\item{object}{an mfa object}
}
\description{
Returns a list of three tables with contributions
}

