% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.r
\docType{methods}
\name{contributions,mfa-method}
\alias{contributions,mfa-method}
\title{Contributions Method for \code{mfa} Object}
\usage{
\S4method{contributions}{mfa}(object)
}
\arguments{
\item{object}{an \code{mfa} object}
}
\value{
A list of tables:

\item{observations}{contribution of each observation to each dimension}

\item{variables}{contribution of each variable to each dimension}

\item{table}{contribution of each table to each dimension}
}
\description{
Returns a list of tables with contribution informations.
}
\examples{

}

