% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{contributions}
\alias{contributions}
\title{Compute contribution}
\usage{
contributions(X, sets, type = 1)
}
\arguments{
\item{X}{- the raw matrix need to be dealt with}

\item{sets}{-  list of vectors indicating the sets of variables (i.e. the blocks)}

\item{type}{- 1: Contribution of an observation to a dimension
2: Contribution of a variable to a dimension
3: Contribution of a table to a dimension.}
}
\value{
contributions : the contribution of required type
}
\description{
This function gets the contribution
}
\examples{
contr <- contributions(X=mtcars,sets=list(1:3,4:6,7:8,9:11),type=1)
}

