% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contributions.R
\name{ContribVar}
\alias{ContribVar}
\title{The contributions of a Variable to a Dimension for a given mfa object.}
\usage{
ContribVar(mfa)
}
\arguments{
\item{mfa}{Object of class 'mfa', which should be output from mfa function.}
}
\value{
A matrix of the contributions of each Variable to each
    Dimension, with rows corresponding to Variables, and columns
    corresponding to Dimensions.
}
\description{
The contributions of a Variable to a Dimension for a given mfa object.
}

