% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{cs_sigmoid}
\alias{cs_sigmoid}
\title{Sigmoid function for activations - renamed to avoid naming conflict}
\usage{
cs_sigmoid(t, phi, k, delta)
}
\value{
Sigmoid function given the parameters
}
\description{
Sigmoid function for activations - renamed to avoid naming conflict
}
\keyword{internal}

