% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{CTD}
\alias{CTD}
\title{Contribution of a Table to a Dimension}
\usage{
CTD(object)
}
\arguments{
\item{object}{an object of class \code{"mfa"}}
}
\value{
object of class \code{"list"}
}
\description{
Outputs values that help interpret how the tables contribute to a dimension. The contribution of a table reﬂects the proportion of the variance of a dimension that can be attributed to this table.
}
\examples{

simulate_data <- matrix(rnorm(200), ncol = 10, byrow = TRUE)
test <- mfa_const(data = simulate_data, sets = list(1:3, 4:5, 6:10), ncomps = 2)
CTD(test)

}

