% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctr_observations.R
\name{ctr_observations}
\alias{ctr_observations}
\title{ctr_observation}
\usage{
ctr_observations(object)
}
\arguments{
\item{object}{a `mfa` object}
}
\value{
a matrix with contribution of an observation to a dimension
}
\description{
ctr_observation
}

