% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctr_tables.R
\name{ctr_tables}
\alias{ctr_tables}
\title{ctr_tables}
\usage{
ctr_tables(object)
}
\arguments{
\item{object}{a `mfa` object}
}
\value{
a matrix with contribution of a table to a dimension
}
\description{
ctr_tables
}

