% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctr_variables.R
\docType{methods}
\name{ctr_variables}
\alias{ctr_variables}
\alias{ctr_variables,mfa-method}
\title{ctr_variables}
\usage{
ctr_variables(object)

\S4method{ctr_variables}{mfa}(object)
}
\arguments{
\item{object}{a `mfa` object}
}
\value{
a matrix with contribution of a variable to a dimension
}
\description{
ctr_variables
}
\section{Methods (by class)}{
\itemize{
\item \code{mfa}: contribution of a variable for mfa
}}

