% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{data_tables}
\alias{data_tables}
\title{data_tables}
\usage{
data_tables(data, sets, center, scale)
}
\arguments{
\item{data}{data set(a dataframe or matrix)}

\item{sets}{a list of vectors indicating sets/blocks of variables, can be character vectors with names or numeric vectors with position of variables in the data table}

\item{center}{logical value or numeric vector of length equal to number of active variables; if numeric vector, each variable has corresponding value subtracted from it; if TRUE, subtract column means}

\item{scale}{logical value or numeric vector of length equal to number of active variables}
}
\value{
a list which contains all tables of the dataset
}
\description{
separate data into individual tables, preprocess and store in a list
}
\examples{

wines <- read.csv("https://raw.githubusercontent.com/ucb-stat243/stat243-fall-2016/master/problem-sets/final-project/data/wines.csv", stringsAsFactors = FALSE)
sets <- list(2:7, 8:13, 14:19, 20:24, 25:30, 31:35, 36:39, 40:45, 46:50, 51:54)
scaling_vec <- apply(subset(wines, select = unlist(sets)), 2, function(x) sqrt(sum((x - mean(x))^2)))
tables <- data_tables(wines, sets, TRUE, scaling_vec)
}

