% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{eigen_table}
\alias{eigen_table}
\title{eigen_table}
\usage{
eigen_table(object)
}
\arguments{
\item{object}{an object of class \code{"mfa"}}
}
\value{
object of class \code{"data.frame"}
}
\description{
Takes the \code{"mfa"} object and returns a table with the singular values (i.e. square root of eigenvalues), the eigenvalues, cumulative, percentage of intertia, cumulative percentage of inertia, for all the extracted components
}
\examples{

simulate_data <- matrix(rnorm(200), ncol = 10, byrow = TRUE)
test <- mfa_const(simulate_data, sets = list(1:3, 4:5, 6:10), ncomps = 2)

eigen_table(test)

}

