% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa_eigenvalue_summary.R
\name{eigensummary}
\alias{eigensummary}
\title{summary of eigenvalues}
\usage{
eigensummary(x)
}
\arguments{
\item{x}{the mfa object created by the mfa function}
}
\value{
a dataframe with singular values, eigenvalues, cumulative, percentage of intertia, cumulative percentage of inertia, for all the extracted components
}
\description{
method for summarizing information about the obtained eigenvalues
}
\examples{

wines <- read.csv("https://raw.githubusercontent.com/ucb-stat243/stat243-fall-2016/master/problem-sets/final-project/data/wines.csv", stringsAsFactors = FALSE)
sets <- list(2:7, 8:13, 14:19, 20:24, 25:30, 31:35, 36:39, 40:45, 46:50, 51:54)
scaling_vec <- apply(subset(wines, select = unlist(sets)), 2, function(x) sqrt(sum((x - mean(x))^2)))
mymfa1 <- mfa(wines, sets, ncomps = 2, T, scaling_vec)

eigensummary(mymfa1)
}

