% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{eigenval}
\alias{eigenval}
\alias{eigenval.mfa}
\title{Eigenvalue summary for object of class \code{"mfa"}}
\usage{
eigenval(x)

\method{eigenval}{mfa}(x)
}
\arguments{
\item{x}{mfa object}
}
\value{
a table with singular values, the eigenvalues, cumulative, percentage of intertia, cumulative percentage of inertia, for all the extracted components.
}
\description{
Creates a table that summarizes information about the obtained eigenvalues
}
\examples{
eigenval(wine)
}

