% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenvalues.r
\docType{methods}
\name{eigenvalues,mfa-method}
\alias{eigenvalues,mfa-method}
\title{Eigenvalues Method for \code{mfa} Object}
\usage{
\S4method{eigenvalues}{mfa}(object)
}
\arguments{
\item{object}{an \code{mfa} object}
}
\value{
a dafa frame contains:

\item{Singular value}{The square root of eigenvalues}

\item{Eigenvalue}{The eigenvalues of \code{mfa} object}

\item{Cumulative}{Cumulative sum of eigenvalues}

\item{precentage  Inertia}{Percentage inertia of eigenvalues}

\item{Cumulative precentage  Inertia}{Cumulative percentage inertia of eigenvalues}
}
\description{
Returns a table with summarizing information about the obtained eigenvalues.
}
\examples{

}

