% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenvalues.r
\name{eigenvalues}
\alias{eigenvalues}
\title{eigenvalues method}
\usage{
eigenvalues(object)
}
\arguments{
\item{object}{an mfa object}
}
\description{
Returns a table with summarizing information about the obtained eigenvalues.
}

