% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FISHub_connect.R
\name{FISHub_connect}
\alias{FISHub_connect}
\title{Connect to the FISHub database}
\usage{
FISHub_connect()
}
\value{
A DBI database connection.
}
\description{
Establishes a connection to the Michigan Fisheries database using Windows integrated authentication.
}
