% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sets.R
\name{format_sets}
\alias{format_sets}
\title{format_sets}
\usage{
format_sets(sets, nGroups)
}
\arguments{
\item{sets}{- the blocks the data should be broken into}

\item{nGroups}{- The number of groups of variables (i.e., number of tables)}
}
\value{
sets - sets as a list, adjusted to be in standard format
}
\description{
private function to put the sets variable into a standard format
}

