% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSVD.R
\name{GSVD}
\alias{GSVD}
\title{Generalized Singular Value Decomposition}
\usage{
GSVD(Data, PLin = NULL, PCol = NULL)
}
\arguments{
\item{Data}{- raw Matrix used for Decomposition}

\item{PLin}{- Vector with weights for lines}

\item{PCol}{- Vector with weights for columns}
}
\value{
List containing:

d - Vector line with the singular values of the result

u - Line eigenvectors

v - Column relative eigenvectors
}
\description{
This function performs Generalized Singular Value Decomposition
}
\examples{
gvsd1 <- GSVD(Data=mtcars)
}

