% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lg.R
\name{LgTable}
\alias{LgTable}
\title{Lg coefficients of subsets of a table}
\usage{
LgTable(data, sets)
}
\arguments{
\item{data}{A data frame or matrix made of multiple tables}

\item{sets}{a list of numeric or character vectors indicating the sets of
variables}
}
\value{
a symmetric matrix containing the Lg coefficients
}

