% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{loadings_plot}
\alias{loadings_plot}
\alias{loadings_plot,mfa-method}
\title{loadings_plot}
\usage{
loadings_plot(x, tablenumber)

\S4method{loadings_plot}{mfa}(x, tablenumber)
}
\arguments{
\item{x}{a `mfa` object}

\item{tablenumber}{the table in the dataset that you wish to display 
loadings for}
}
\value{
A plot of the variable loadings
}
\description{
loadings_plot
}
\section{Methods (by class)}{
\itemize{
\item \code{mfa}: a plot of the variable loadings
}}

