% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Log sum of exponentials}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{Vector of quantities}
}
\value{
Log sum of exponentials in a numerically stable manner.
}
\description{
Log sum of exponentials
}
\keyword{internal}
