% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{map_branch}
\alias{map_branch}
\title{Find the MAP branch and uncertainty}
\usage{
map_branch(g)
}
\arguments{
\item{g}{The trace slot from an mfa fit}
}
\value{
A data frame with a column corresponding
to the MAP branch and a further column corresponding
to the proportion of samples assigned to that branch
(a measure of uncertainty).
}
\description{
Find the MAP branch and uncertainty
}
\keyword{internal}
