% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R, R/zi.R
\name{mcmcify}
\alias{mcmcify}
\title{Turn a matrix's columns into informative names}
\usage{
mcmcify(m, name)

mcmcify(m, name)
}
\description{
Turn a matrix's columns into informative names

Turn a matrix's columns into informative names
}

