% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{mfa}
\alias{mfa}
\title{mfa}
\usage{
mfa(data, sets, ncomps = NULL, center = TRUE, scale = TRUE)
}
\arguments{
\item{data}{data set (matrix or data frame). Dimensions must be observations X variables.}

\item{sets}{list of vectors indicating the sets of variables (i.e. the mfa blocks).}

\item{ncomps}{integer indicating how many components (i.e. factors) should be used in the dimensionality reduction.}

\item{center}{either a logical value or a numeric vector of length equal to the number of active
variables in the analysis. (passed to the \code{scale} function)}

\item{scale}{either a logical value or a numeric vector of length equal to the number of active
variables in the analysis (passed to the \code{scale} function)}
}
\value{
The returned value is an object of class "mfa" with the following elements: \cr
\code{eigenvalues} a vector containing eigenvalues for the final factors \cr
\code{cfs} a matrix of common factor scores \cr
\code{pfs} a list of partial factor scores for each table \cr
\code{loadings} a matrix of loadings \cr
\code{a_weights} alpha weight for each table \cr
}
\description{
Multiple Factor Analysis
}
\examples{
data(wines)
wines = wines[,2:54]
sets = list(1:6,7:12,13:18,19:23,24:29,30:34,35:38,39:44,45:49,50:53)
obj = mfa(wines, sets, ncomps = 2, center = TRUE, scale = TRUE)
}

