% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{mfa_const}
\alias{mfa_const}
\title{mfa_const}
\usage{
mfa_const(data, sets, ncomps = NULL, center = TRUE, scale = TRUE)
}
\arguments{
\item{data}{data set (matrix or data frame)}

\item{sets}{list of vectors indicating the sets of variables (i.e. the blocks)}

\item{ncomps}{integer indicating how many number of components (i.e. factors) are to be extracted}

\item{center}{logical value}

\item{scale}{logical value}
}
\value{
an object of class mfa
}
\description{
Creates an object of class \code{"mfa"}
}
\examples{

simulate_data <- matrix(rnorm(200), ncol = 10, byrow = TRUE)
test <- mfa_const(data = simulate_data, sets = list(1:3, 4:5, 6:10), ncomps = 2)

}

