% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{mfa_gen}
\alias{mfa_gen}
\title{Construct mfa}
\usage{
mfa_gen(data, sets, ncomps = NULL, center = TRUE, scale = TRUE)
}
\arguments{
\item{data}{dataset on which MFA is conducted}

\item{sets}{list of vectors indicating the sets of variables (i.e. the blocks)}

\item{ncomps}{integer indicating how many number of components (i.e. factors) are to be extracted}

\item{center}{either a logical value or a numeric vector of length equal to the number of active variables in the analysis}

\item{scale}{either a logical value or a numeric vector of length equal to the number of active variables in the analysis}
}
\value{
an object of class mfa
}
\description{
Creates an object of class \code{"mfa"}
}
\examples{
url="https://raw.githubusercontent.com/ucb-stat243/stat243-fall-2016/master/problem-sets/final-project/data/wines.csv"
data=read.csv(file=url, header = TRUE, sep = ",")
datanum <- data[,-1]

# a list of numeric vectors with the position of the active variables in the data table.
sets <- list(1:6, 7:12, 13:18, 19:23, 24:29, 30:34, 35:38, 39:44, 45:49, 50:53)
# create object 'wine' of class mfa
wine <- mfa_gen(data = datanum, sets = sets, ncomps = 2)
}

