% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_dim.R
\name{obs_dim}
\alias{obs_dim}
\title{Contribution: observation to dimension}
\usage{
obs_dim(obj)
}
\arguments{
\item{obj}{is an mfa object}
}
\description{
Calculates the constribution of an observation to a dimension.
}

