% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{pfs_plot}
\alias{pfs_plot}
\alias{pfs_plot,mfa-method}
\title{pfs_plot}
\usage{
pfs_plot(x, tablenumber)

\S4method{pfs_plot}{mfa}(x, tablenumber)
}
\arguments{
\item{x}{a `mfa` object}

\item{tablenumber}{The number of the table to display partial 
factor scores for.}
}
\value{
A plot of the partial factor scores for a table.
}
\description{
pfs_plot
}
\section{Methods (by class)}{
\itemize{
\item \code{mfa}: plot of the partial factor scores for a table
}}

