% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compromise.R
\name{plot compromise}
\alias{plot compromise}
\alias{plot_compromise.mfa}
\title{Plot Compromise}
\usage{
\method{plot_compromise}{mfa}(x, dim1 = 1, dim2 = 2, sz = 2,
  obsnames = NULL, textcolor = "black")
}
\arguments{
\item{x}{An object of class mfa}

\item{dim1}{dimension for x axis, default 1st component}

\item{dim2}{dimension for y axis, default 2nd component}

\item{sz}{size of text label in scatter plot}

\item{obsnames}{input labels for your observations if desired}

\item{textcolor}{color of text labeling points, can be one color or a vector of length num obs colors}
}
\description{
plots the compromise, i.e. the common factor scores, for 2 given dimensions of an mfa table
}

