% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mfa.R
\name{plot loadings}
\alias{plot loadings}
\alias{plot_loading}
\title{Plot Loadings}
\usage{
plot_loading(x, table = 1, dim1 = 1, dim2 = 2, sz = 2)
}
\arguments{
\item{x}{An object of class mfa}

\item{table}{which table to plot (integer 1:K)}

\item{dim1}{dimension for x axis, default 1st component}

\item{dim2}{dimension for y axis, default 2nd component}

\item{sz}{size of point on scatter plot}
}
\description{
plots loadings for object of class mfa
}

