% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loading.R
\name{plot matrix loadings}
\alias{plot matrix loadings}
\alias{plot_loading.mfa}
\title{Plot Loadings}
\usage{
\method{plot_loading}{mfa}(x, table = 1, dim1 = 1, dim2 = 2, sz = 2,
  varnames = NULL)
}
\arguments{
\item{x}{An object of class mfa}

\item{table}{which table to plot (integer 1:K)}

\item{dim1}{dimension for x axis, default 1st component}

\item{dim2}{dimension for y axis, default 2nd component}

\item{sz}{size of point on scatter plot}

\item{varnames}{variable labels}
}
\description{
plots matrix loadings
}

