% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_partial_fac.R
\name{plot partial fac}
\alias{plot partial fac}
\alias{plot_partial_fac}
\alias{plot_partial_fac.mfa}
\title{Plot Partial Fac}
\usage{
plot_partial_fac(obj, table = 1, dim1 = 1, dim2 = 2, sz = 2)

\method{plot_partial_fac}{mfa}(x, table = 1, dim1 = 1, dim2 = 2, sz = 2)
}
\arguments{
\item{obj}{An object of class mfa}

\item{table}{which table to plot (integer 1:K)}

\item{dim1}{dimension for x axis, default 1st component}

\item{dim2}{dimension for y axis, default 2nd component}

\item{sz}{size of point on scatter plot}

\item{x}{An object of class mfa}
}
\description{
plots partial factor scores

plots partial factor scores
}

