% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{plot.bootstrap}
\alias{plot.bootstrap}
\title{Plot method for 'bootstrap' object}
\usage{
\method{plot}{bootstrap}(x, ...)
}
\arguments{
\item{x}{- bootstrap-object}

\item{\dots}{further arguments ignored}
}
\value{
plot method for "bootstrap" object
}
\description{
This function defines a plot method for "bootstrap" object
}

