% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa_plot_method.R
\name{plot.mfa}
\alias{plot.mfa}
\title{Plot of object mfa}
\usage{
\method{plot}{mfa}(mfa, type, d1 = 1, d2 = 2, X = 1,
  loading_labels = NULL, cex = 1, ...)
}
\arguments{
\item{mfa}{an object of class \code{"mfa"}}

\item{type}{indicating which plot to output. 1: factor score plot; 2:factor score and partial loadings; 3: produce factor score and partial loadings on all components; 4: partial factor score; 5: partial loadings}

\item{d1}{indicating the x-axis}

\item{d2}{indicating the y-axis}

\item{X}{indicating the target table to produce a plot}

\item{loading_labels}{the labels for partial loadings}

\item{cex}{parameter for controlling the size of title, lengends, labels, and points}

\item{\dots}{arguments to be passed to/from other methods}
}
\description{
Plots factor score, partial factor score and partial loadings
}
\examples{

wines <- read.csv("https://raw.githubusercontent.com/ucb-stat243/stat243-fall-2016/master/problem-sets/final-project/data/wines.csv", stringsAsFactors = FALSE)
sets <- list(2:7, 8:13, 14:19, 20:24, 25:30, 31:35, 36:39, 40:45, 46:50, 51:54)
scaling_vec <- apply(subset(wines, select = unlist(sets)), 2, function(x) sqrt(sum((x - mean(x))^2)))
mymfa <- mfa(wines, sets, ncomps = 2, T, scaling_vec)

plot(mymfa, type = 1)
plot(mymfa, type = 2, X = 1)
plot(mymfa, type = 2, X = 1, loading_labels = c("Cat Pee", "Passion Fruit", "Green Pepper", "Mineral", "Smoky", "Citrus"))
plot(mymfa, type = 3, loading_labels = NULL)
plot(mymfa, type = 4, X = 1)
plot(mymfa, type = 5, X = 1)

}

