% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{plot_chi}
\alias{plot_chi}
\title{Plot posterior precision parameters}
\usage{
plot_chi(m, nfeatures = m$G)
}
\arguments{
\item{m}{A fit returned from \code{mfa}}

\item{nfeatures}{Top number of}
}
\value{
A \code{ggplot2} bar-plot showing the map
estimates of \eqn{\chi^{-1}}
}
\description{
Plot posterior precision parameters
}

