% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compromise.R
\name{plot_compromise}
\alias{plot_compromise}
\title{plot_compromise}
\usage{
plot_compromise(mfa, dims = 1:2, title = "Common Factor Scores")
}
\arguments{
\item{mfa}{- the mfa object}

\item{dims}{- Numeric vector of two values signifying dimensions to plot}

\item{title}{- chart title}
}
\value{
a plot with the projection of each observation onto the first two
        extracted components
}
\description{
Plot the compromise scores for the first two extracted dimensions
}

