% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{plot_dropout_relationship}
\alias{plot_dropout_relationship}
\title{Plot the dropout relationship}
\usage{
plot_dropout_relationship(y, lambda = empirical_lambda(y))
}
\arguments{
\item{y}{The input data matrix}

\item{lambda}{The estimated value of lambda}
}
\value{
A \code{ggplot2} plot showing the estimated dropout relationship
}
\description{
Plot the dropout relationship
}
\examples{
synth <- create_synthetic(C = 20, G = 5, zero_negative = TRUE, model_dropout = TRUE)
lambda <- empirical_lambda(synth$X)
plot_dropout_relationship(synth$X, lambda)
}

