% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{plot_mfa_autocorr}
\alias{plot_mfa_autocorr}
\title{Plot MFA autocorrelation}
\usage{
plot_mfa_autocorr(m)
}
\arguments{
\item{m}{A fit returned from \code{mfa}}
}
\value{
A \code{ggplot2} plot returned by the \code{ggmcmc} package plotting
the autocorrelation of the posterior log-likelihood.
}
\description{
Plots the autocorrelation of the posterior log-likelihood.
}
\examples{
synth <- create_synthetic(C = 20, G = 5)
m <- mfa(synth$X)
plot_mfa_autocorr(m)
}

