% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{plot_mfa_trace}
\alias{plot_mfa_trace}
\title{Plot MFA trace}
\usage{
plot_mfa_trace(m)
}
\arguments{
\item{m}{A fit returned from \code{mfa}}
}
\value{
A \code{ggplot2} plot plotting
the trace of the posterior log-likelihood.
}
\description{
Plots the trace of the posterior log-likelihood.
}
\examples{
synth <- create_synthetic(C = 20, G = 5)
m <- mfa(synth$X)
plot_mfa_trace(m)
}
