% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_partialfs}
\alias{plot_partialfs}
\title{Plot of partial factor scores and variable loadings for two selected dimensions}
\usage{
plot_partialfs(data, sets, center = TRUE, scale = TRUE, dim1, dim2, labels)
}
\arguments{
\item{data}{- data set}

\item{sets}{- list of vectors indicating the sets of variables}

\item{center}{- logical argument, whether to center}

\item{scale}{- logical arugment, whether to scale}

\item{dim1}{- integer, selected dimension to plot on x-axis}

\item{dim2}{- integer, selected dimension to plot on y-axis}

\item{labels}{- label for each row}
}
\value{
plots of partial factor scores and variable loadings for the two selected axes
}
\description{
This function generates plots of partial factor scores and variable loadings for two selected dimensions
}
\examples{
plot_partialfs(mtcars[c(1:2,4:5,8:9,20:21),],sets=list(1:3,4:6,7:8,9:11),dim1=1,dim2=2,labels=c("Maz1","Maz2","Hor1","Hor2","Merc1","Merc2","Toyota1","Toyota2"))
}

