% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{plot_vload}
\alias{plot_vload}
\title{Plot of Variable Loadings}
\usage{
plot_vload(dimension1, dimension2,
  rownames_vec = as.character(1:length(dimension1)))
}
\arguments{
\item{dimension1}{vector of class \code{"numeric"}}

\item{dimension2}{vector of class \code{"numeric"}}

\item{rownames_vec}{vector of row labels}
}
\value{
plot of variable loadings
}
\description{
Takes two dimensions of variable loadings and returns a plot of the two dimensions
}
\examples{

dimension1 <- rnorm(20)
dimension2 <- rnorm(20)
plot_compromise(dimension1, dimension2)

}

