% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotFactorScores}
\alias{PlotFactorScores}
\title{Plot Factor Score}
\usage{
PlotFactorScores(mfa, group = NULL, dim = c(1, 2))
}
\arguments{
\item{mfa}{an object of class "mfa"}

\item{group}{a factor vector, indicating which group each row
belongs to}

\item{dim}{a integer vector indicating the dimensions to plot, the
first element being x and second being y}
}

