% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotLoadings}
\alias{PlotLoadings}
\title{Plot Variable Loadings}
\usage{
PlotLoadings(mfa, var.group = NULL, label = NULL, dim = c(1, 2))
}
\arguments{
\item{mfa}{an object of class "mfa"}

\item{var.group}{a factor vector, indicating which group each row
(variable) belongs to}

\item{label}{a character vector, will be used to plot the points,
default to row numbers when NULL}

\item{dim}{a integer vector indicating the dimensions to plot, the
first element being x and second being y}
}

