% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotPartialFactorScores}
\alias{PlotPartialFactorScores}
\title{Plot Partial Factor Scores}
\usage{
PlotPartialFactorScores(mfa, k, group = NULL, label = NULL, dim = c(1, 2))
}
\arguments{
\item{mfa}{an object of class "mfa"}

\item{k}{the partial factor score of which table}

\item{group}{a factor vector, indicating which group each row
belongs to}

\item{label}{a character vector, will be used to plot the points,
default to row numbers when NULL}

\item{dim}{a integer vector indicating the dimensions to plot, the
first element being x and second being y}
}

