% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\docType{methods}
\name{print,mfa-method}
\alias{print,mfa-method}
\title{Pringt Method for \code{mfa} Object}
\usage{
\S4method{print}{mfa}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"mfa"}}

\item{...}{more arguments}
}
\value{
Some summary information about the \code{mfa} object:

The number of components

The eigenvalue of the first component

The eigenvalue of the second component
}
\description{
Pringting an \code{mfa} object.
}
\details{
Pringting MFAs
}
\examples{
test<-mfa(wine,sets=list(1:6,7:12,13:18,19:23,24:29,30:34,35:38,39:44,45:49,50:53))
print(test)
}

