% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-mfa.R
\name{print mfa}
\alias{print mfa}
\alias{print.mfa}
\title{Print mfa}
\usage{
\method{print}{mfa}(x)
}
\arguments{
\item{x}{an object of class mfa}
}
\description{
prints information about an mfa object. note when using this method,
the user will be prompted to input which table she would like to see
for partial factor scores and for variable loadings
}

