% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa_print_method.R
\name{print.mfa}
\alias{print.mfa}
\title{print method of object mfa}
\usage{
\method{print}{mfa}(x)
}
\arguments{
\item{x}{an object of class \code{"mfa"}}
}
\description{
print the basic information of a mfa object
}
\examples{

wines <- read.csv("https://raw.githubusercontent.com/ucb-stat243/stat243-fall-2016/master/problem-sets/final-project/data/wines.csv", stringsAsFactors = FALSE)
sets <- list(2:7, 8:13, 14:19, 20:24, 25:30, 31:35, 36:39, 40:45, 46:50, 51:54)
scaling_vec <- apply(subset(wines, select = unlist(sets)), 2, function(x) sqrt(sum((x - mean(x))^2)))
mymfa1 <- mfa(wines, sets, ncomps = 2, T, scaling_vec)

mymfa1
}

