% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{print,mfa-method}
\alias{print,mfa-method}
\title{print}
\usage{
\S4method{print}{mfa}(x, compromise = TRUE, pfs = TRUE, loadings = TRUE,
  tablenumber = 1)
}
\arguments{
\item{x}{a `mfa` object}

\item{compromise}{boolean; output a plot of the first two components of compromise factor score if TRUE}

\item{pfs}{boolean; output a plot of the first two components of partial factor score for table with specified `tablenumber`if TRUE}

\item{loadings}{boolean; output a plot of the first two components of loadings for table with specified `tablenumber` if TRUE}

\item{tablenumber}{table number, used for partial factor score or loadings plot}
}
\value{
A visualization table; also plots if any of parameters `compromise`, `pfs`,
`loadings` are set to be TRUE
}
\description{
print
}

