% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV_table.R
\name{RV_table}
\alias{RV_table}
\title{RV_table}
\usage{
RV_table(dataset, sets)
}
\arguments{
\item{dataset}{a data frame or matrix}

\item{sets}{list of vectors indicating the sets of variables (i.e. the blocks).}
}
\value{
a matrix
}
\description{
To evaluate the similarities between each pair of tables among all tables in the dataset
}

