% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\name{sumeigen}
\alias{sumeigen}
\title{summaries of eigenvalues}
\usage{
sumeigen(Mfa)
}
\arguments{
\item{Mfa}{a `mfa` class object}
}
\value{
a table
}
\description{
returns a table with the singular values, the eigenvalues, cumulative, percentage of intertia, cumulative percentage of inertia, for all the extracted components.
}

