% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mfa}
\alias{summary.mfa}
\title{Print summary of eigen values of MFA}
\usage{
\method{summary}{mfa}(MFA, ncomps = NULL)
}
\arguments{
\item{MFA}{- a "mfa" object}

\item{ncomps}{- the number of eigen values -default}
}
\value{
List containing:

Singular values

Eigenvalues

% cumulative eigenvalues

% Inertia

% cumulative Inertia
}
\description{
This function gets summary the eigens value of MFA
}

