% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_eigenvalues.R
\name{summary_eigenvalues}
\alias{summary_eigenvalues}
\title{summary_eigenvalues}
\usage{
summary_eigenvalues(mfa)
}
\arguments{
\item{mfa}{- the mfa object to summarize}
}
\description{
provide a print-out summary of the eigenvalues of an mfa object
}

