% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_dim.R
\name{table_dim.mfa}
\alias{table_dim.mfa}
\title{Contribution: table to dimension}
\usage{
\method{table_dim}{mfa}(x)
}
\arguments{
\item{x}{is an mfa object}
}
\value{
a matrix k-by-l where k is the tables and l is the dimensions
}
\description{
Calculates the contributions of a table to the dimension
}

