% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_dim.R
\name{var_dim.mfa}
\alias{var_dim.mfa}
\title{Contribution: variable to dimension}
\usage{
\method{var_dim}{mfa}(x)
}
\arguments{
\item{x}{is an mfa object}
}
\value{
ctr_var a n-by-m matrix where n is the variables and m is the dimensions
}
\description{
Calculates the constribution of a variable to a dimension
}

