% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{wine}
\alias{wine}
\title{A dataset containing Wine score of 12 wines from 10 wine accessors.}
\format{An object of class \code{matrix} with 12 rows and 53 columns.}
\usage{
wine
}
\description{
A dataset containing Wine score of 12 wines from 10 wine accessors.
}
\keyword{datasets}

