% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wines}
\alias{wines}
\title{Wine tasting experiment data}
\format{A data frame with 12 rows and 58 variables}
\source{
\url{"https://raw.githubusercontent.com/ucb-stat243/stat243-fall-2016/master/problem-sets/final-project/data/wines.csv"}
}
\usage{
wines
}
\description{
The working example consists of a (fictitious) wine tasting experiment. 
The data concerns 12 wines made from Sauvignon Blanc grapes coming from 
three wine regions (4 wines from each region): New Zealand, France, and 
Canada. Ten expert assessors were asked to evaluate these wines. The 
assessors were asked:
\enumerate{
  \item to evaluate the wines on 9-point rating scales, using four variables 
  considered as standard for the evaluation of these wines: cat-pee, 
  passion-fruit, green pepper, and mineral.
  \item if they felt the need, to add some variables of their own: some 
  assessors choose none, some choose one or two more variables.
}
}
\keyword{datasets}

