% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap for object of class \code{"mfa"}}
\usage{
bootstrap(L, dataset, userset, ncomps)
}
\arguments{
\item{L}{size of the bootstrap sample}

\item{dataset}{the grand table}

\item{userset}{list of vectors to partition the grand table into subtables}

\item{ncomps}{integer indicating how many number of components (i.e. factors) are to be extracted}
}
\value{
matrix of bootstrap ratios
}
\description{
Perform bootstrap in order to get the matrix of bootstrap ratio. The ratios are used to find the observations that reliably contributes to a given component
}
\examples{
sets <- list(1:6, 7:12, 13:18, 19:23, 24:29, 30:34, 35:38, 39:44, 45:49, 50:53)
bootstrap(1000,sets,datanum,2)
}

