% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{bootstrap}
\usage{
bootstrap(mfa, B = 100, seed = 0)
}
\arguments{
\item{mfa}{an mfa object}

\item{B}{the number of bootstrap samples to run}

\item{seed}{a seed for the random number generator}
}
\value{
a copy of the mfa object with the bootstrap as an
        attribute.
}
\description{
Perform a boostrap approximation for the compromise of the
given mfa object
}

