% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Simulate bootstrap sampling}
\usage{
bootstrap(dataset, sets, ncomps = NULL, center = TRUE, scale = TRUE, k, B)
}
\arguments{
\item{dataset}{- data set}

\item{sets}{- list of vectors indicating the sets of variables}

\item{ncomps}{- integer, how many number of components}

\item{center}{- logical argument, whether to center}

\item{scale}{- logical arugment, whether to scale}

\item{k}{- integer, size of sampling for each bootstrap variable}

\item{B}{- integer, size of bootstrap sampling}
}
\value{
List containing:

Mean  - List of bootstrap means for each axis

Variance  - List of bootstrap variance for each axis

Ratio  - List of bootstrap ratio for each axis

Size  - Size of bootstrap samples

Dimension - Dimension of bootstrap samples (number of axis)
}
\description{
This function simulates bootstrap sampling of partial factor scores
}
\examples{
b.sample <- bootstrap(dataset=mtcars,sets=list(1:3,4:6,7:8,9:11), ncomps=NULL, center=TRUE, scale=TRUE, k=10,B=100)
}

