% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_factorscores.R
\name{bootstrap_factorscores.mfa}
\alias{bootstrap_factorscores.mfa}
\title{Bootstrapping Factor Scores}
\usage{
\method{bootstrap_factorscores}{mfa}(x)
}
\arguments{
\item{x}{is an mfa object}
}
\value{
returns a 3-by-1 list containing the mean bootstrapped factor scores, the variance of the bootstrapped factor scores and the associated T star statistic.
}
\description{
Calculates the bootstrap confidence intervals by sampling with replace of factor scores
}
\examples{
b bootstrap factor score
from the set of tables. This approach also computes bootstrap ratios for
each dimension

}

