% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.r
\name{contributions}
\alias{contributions}
\title{Contributions}
\usage{
contributions(object)
}
\arguments{
\item{object}{an \code{mfa} object}
}
\value{
A list of tables:

\item{observations}{contribution of each observation to each dimension}

\item{variables}{contribution of each variable to each dimension}

\item{table}{contribution of each table to each dimension}
}
\description{
Returns a list of tables with contribution informations.
}
\examples{
ndatas<-apply(wine,2,function(x){ (x-mean(x))/norm(x-mean(x),type="2")})
test<-mfa(ndatas,sets=list(1:6,7:12,13:18,19:23,24:29,30:34,35:38,39:44,45:49,50:53),
          center=FALSE,scale=FALSE)
contributions(test)
}

