% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{create_synthetic}
\alias{create_synthetic}
\title{Create synthetic data}
\usage{
create_synthetic(C = 100, G = 40, p_transient = 0, zero_negative = TRUE,
  model_dropout = FALSE, lambda = 1)
}
\description{
Create synthetic data
}

