% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctr_observations.R
\docType{methods}
\name{ctr_observations}
\alias{ctr_observations}
\alias{ctr_observations,mfa-method}
\title{ctr_observations}
\usage{
ctr_observations(object)

\S4method{ctr_observations}{mfa}(object)
}
\arguments{
\item{object}{a `mfa` object}
}
\value{
a matrix with contribution of an observation to a dimension
}
\description{
ctr_observations
}
\section{Methods (by class)}{
\itemize{
\item \code{mfa}: contribution of an observation for mfa
}}

