% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctr_tables.R
\docType{methods}
\name{ctr_tables}
\alias{ctr_tables}
\alias{ctr_tables,mfa-method}
\title{ctr_tables}
\usage{
ctr_tables(object)

\S4method{ctr_tables}{mfa}(object)
}
\arguments{
\item{object}{a `mfa` object}
}
\value{
a matrix with contribution of a table to a dimension
}
\description{
ctr_tables
}
\section{Methods (by class)}{
\itemize{
\item \code{mfa}: contribution of a table for mfa
}}
\examples{
data(wines)
wines = wines[,2:54]
sets = list(1:6,7:12,13:18,19:23,24:29,30:34,35:38,39:44,45:49,50:53)
obj = mfa(wines, sets, ncomps = 2, center = TRUE, scale = TRUE)
ctr_tables(obj)
}

