% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctr_tables.R
\docType{methods}
\name{ctr_tables}
\alias{ctr_tables}
\alias{ctr_tables,mfa-method}
\title{ctr_tables}
\usage{
ctr_tables(object)

\S4method{ctr_tables}{mfa}(object)
}
\arguments{
\item{object}{a `mfa` object}
}
\value{
a matrix with contribution of a table to a dimension
}
\description{
ctr_tables
}
\section{Methods (by class)}{
\itemize{
\item \code{mfa}: contribution of a table for mfa
}}

