% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctr_variables.R
\name{ctr_variables}
\alias{ctr_variables}
\title{ctr_variables}
\usage{
ctr_variables(object)
}
\arguments{
\item{object}{a `mfa` object}
}
\value{
a matrix with contribution of a variable to a dimension
}
\description{
ctr_variables
}

