% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctr_variables.R
\docType{methods}
\name{ctr_variables}
\alias{ctr_variables}
\alias{ctr_variables,mfa-method}
\title{ctr_variables}
\usage{
ctr_variables(object)

\S4method{ctr_variables}{mfa}(object)
}
\arguments{
\item{object}{a `mfa` object}
}
\value{
a matrix with contribution of a variable to a dimension
}
\description{
ctr_variables
}
\section{Methods (by class)}{
\itemize{
\item \code{mfa}: contribution of a variable for mfa
}}
\examples{
data(wines)
wines = wines[,2:54]
sets = list(1:6,7:12,13:18,19:23,24:29,30:34,35:38,39:44,45:49,50:53)
obj = mfa(wines, sets, ncomps = 2, center = TRUE, scale = TRUE)
ctr_variables(obj)
}

