% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{eigenvalue table}
\alias{eigenvalue table}
\alias{eigenvalueTable}
\alias{eigenvalueTable.mfa}
\title{Eigenvalue Table}
\usage{
eigenvalueTable(x)

\method{eigenvalueTable}{mfa}(obj)
}
\arguments{
\item{x}{An object of class mfa}

\item{obj}{An object of class mfa}
}
\description{
eigenvalue table from mfa obj

eigenvalue table from mfa obj
}

