% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenvalues.r
\name{eigenvalues}
\alias{eigenvalues}
\title{Eigenvalues}
\usage{
eigenvalues(object)
}
\arguments{
\item{object}{an \code{mfa} object}
}
\value{
a dafa frame contains:

\item{Singular value}{The square root of eigenvalues}

\item{Eigenvalue}{The eigenvalues of \code{mfa} object}

\item{Cumulative}{Cumulative sum of eigenvalues}

\item{precentage Inertia}{Percentage inertia of eigenvalues}

\item{Cumulative percentage Inertia}{Cumulative percentage inertia of eigenvalues}
}
\description{
Returns a table with summarizing information about the obtained eigenvalues.
}
\examples{
test<-mfa(wine,sets=list(1:6,7:12,13:18,19:23,24:29,30:34,35:38,39:44,45:49,50:53))
eigenvalues(test)

}

