% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{empirical_lambda}
\alias{empirical_lambda}
\title{Estimate the dropout parameter}
\usage{
empirical_lambda(y, lower_limit = 0)
}
\arguments{
\item{y}{A cell-by-gene expression matrix}

\item{lower_limit}{The limit below which expression counts as 'dropout'}
}
\value{
The estimated lambda
}
\description{
Estimate the dropout parameter
}
\examples{
synth <- create_synthetic(C = 20, G = 5, zero_negative = TRUE, model_dropout = TRUE)
lambda <- empirical_lambda(synth$X)
}
