% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{Lg}
\alias{Lg}
\title{Lg Coefficient}
\usage{
Lg(table1, table2)
}
\arguments{
\item{table1}{object of class \code{"matrix"} with same number of rows as table 2}

\item{table2}{object of class \code{"matrix"} with same number of rows as table 1}
}
\value{
object of class \code{"numeric"}
}
\description{
Computes the Lg coefficient between two tables
}
\examples{

table1 <- matrix(rnorm(100), nrow = 10)
table2 <- matrix(rnorm(50), nrow = 10)
Lg(table1, table2)

}

