% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lg.R
\name{lg}
\alias{lg}
\title{Lg Coefficient - lg}
\usage{
lg(table1, table2)
}
\arguments{
\item{table1}{is a list or matrix of the  first input table to be included in the analysis}

\item{table2}{is a list or matrix of the second input table to be included in the analysis}
}
\description{
Calculates Lg coefficient between two tables
}
\examples{
filename = system.file("extdata", "wines.csv", package = "MFA")
d = read.csv(filename, header=TRUE, check.names=FALSE)
table1 = d[,seq(2,7)]
table2 = d[,seq(8,13)]
lg(table1,table2)

}

