% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{lg}
\alias{lg}
\title{Lg Coefficient - lg}
\usage{
lg(table1, table2)
}
\arguments{
\item{x}{An object of class mfa
example lg(table1, table2)}
}
\description{
Calculates Lg coefficient between two tables.
}
\details{
details
}

