% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lg.R
\name{Lg}
\alias{Lg}
\title{Lg}
\usage{
Lg(table1, table2)
}
\arguments{
\item{table1, }{a matrix}

\item{table2, }{a matrix}
}
\value{
the Lg coefficient (scalar)
}
\description{
Computes the Lg coefficient between two tables
}
\examples{

wines <- read.csv("https://raw.githubusercontent.com/ucb-stat243/stat243-fall-2016/master/problem-sets/final-project/data/wines.csv", stringsAsFactors = FALSE)
sets <- list(2:7, 8:13, 14:19, 20:24, 25:30, 31:35, 36:39, 40:45, 46:50, 51:54)
scaling_vec <- apply(subset(wines, select = unlist(sets)), 2, function(x) sqrt(sum((x - mean(x))^2)))
tables <- data_tables(wines, sets, TRUE, scaling_vec)

Lgtable1_table2 <- Lg(tables[[1]], tables[[2]])
}

