% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lg.R
\name{Lg}
\alias{Lg}
\title{Lg}
\usage{
Lg(table1, table2)
}
\arguments{
\item{table1}{- first numeric data.frame/matrix}

\item{table2}{- second numeric data.frame/matrix}
}
\description{
Function to compute the Lg coefficient between two numeric tables
}

