% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG.r
\name{LG}
\alias{LG}
\title{Lg Coefficients}
\usage{
LG(table1, table2)
}
\arguments{
\item{table1}{a normalized data matrix or data.frame}

\item{table2}{a normalized data matrix or data.frame}
}
\value{
a value of Lg coefficient between two tables
}
\description{
Return a value of Lg coefficient between two tables.
}
\examples{
# default
ndatas<-apply(wine,2,function(x){ (x-mean(x))/norm(x-mean(x),type="2")})
x1 <- ndatas[,1:6]
x2 <- ndatas[,7:12]
LG(x1,x2)
}

