% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lg.R
\name{Lg}
\alias{Lg}
\title{Compute Lg coefficient between 2 tables}
\usage{
Lg(MFA, table1, table2, order1, order2)
}
\arguments{
\item{MFA}{- a "mfa" object}

\item{table1}{- matrix of the first table}

\item{table2}{- matrix of the second table}

\item{order1}{- the number of table1 in order}

\item{order2}{- the number of table2 in order}
}
\value{
lg - lg coefficients of two tables
}
\description{
This function gets the Lg coefficients to study the Between-Table Structure
}
\examples{
mfa1 <- mfa(data=mtcars,sets=list(1:3,4:6,7:8,9:11), ncomps = NULL, center = TRUE, scale = TRUE)
table_1 <- mfa1$MatrixZ[,1:3]
table_2 <- mfa1$MatrixZ[,4:6]
lg <- Lg(MFA=mfa1,table1=table_1,table2=table_2,order1=1,order2=2)
}

