% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG.r
\name{LG}
\alias{LG}
\title{Lg Coefficients}
\usage{
LG(table1, table2)
}
\arguments{
\item{table1}{a normalized data matrix or data.frame}

\item{table2}{a normalized data matrix or data.frame}
}
\value{
a value of Lg coefficient between two tables
}
\description{
Return a value of Lg coefficient between two tables.
}
\examples{
# default
x1 <- scale(wine_data[,1:6])
x2 <- scale(wine_data[,7:12])
LG(x1,x2)
}

