% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{Lg_table}
\alias{Lg_table}
\title{Lg Table}
\usage{
Lg_table(dataset, sets)
}
\arguments{
\item{dataset}{object of class \code{"matrix"} or \code{"data.frame"}}

\item{sets}{object of class \code{"list"}}
}
\value{
object of class \code{"matrix"}
}
\description{
Computes a table of Lg coefficients.
}
\examples{

simulate_data <- matrix(rnorm(200), ncol = 10, byrow = TRUE)
Lg_table(simulate_data, sets = list(1:3, 4:5, 6:10))

}

