% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{lg_table}
\alias{lg_table}
\title{Lg Coefficient Table - lg_table}
\usage{
lg_table(dataset, sets)
}
\arguments{
\item{x}{An object of class mfa
example lg_table(dataset, sets = list(1:3, 4:5, 6:10))
         returns a 3-by-3 symmetric matrix.}
}
\description{
Calculates Lg coefficient table between multiple sets
}
\details{
details
}

