% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lg_table.R
\name{Lg_table}
\alias{Lg_table}
\title{Lg_table}
\usage{
Lg_table(mfa, sets = NULL)
}
\arguments{
\item{mfa}{- the mfa object containing the data}

\item{sets}{- a numeric list indicating the groups of columns
that correspond to each table}
}
\description{
computes a matrix of Lg scores for all the tables used in an
multi-factor analysis.
}

