% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_table.r
\name{LG_table}
\alias{LG_table}
\title{Lg Coefficients Table}
\usage{
LG_table(dataset, sets)
}
\arguments{
\item{dataset}{a normalized dataframe or matrix}

\item{sets}{list of vector contains vector of indices or variable names of each group}
}
\value{
a table of Lg coefficients
}
\description{
Return a table of Lg coefficients between any two subsets of a normalized dataset.
}
\examples{
# numeric sets
ndatas<-apply(wine,2,function(x){ (x-mean(x))/norm(x-mean(x),type="2")})
LG_table(ndatas,sets=list(1:6,7:12,13:18,19:23,24:29,30:34,35:38,39:44,45:49,50:53))

# character sets
ndatas<-apply(wine,2,function(x){ (x-mean(x))/norm(x-mean(x),type="2")})
# Use the first and last variable names of each group:
LG_table(ndatas,sets=list(c("V1.G1","V6.G1"),c("V1.G2","V8.G2"),c("V1.G3","V10.G3"),
                           c("V1.G4","V8.G4"),c("V1.G5","V12.G5"),c("V1.G6","V13.G6"),
                           c("V1.G7","V4.G7"),c("V1.G8","V5.G8"),c("V1.G9","V15.G9"),
                           c("V1.G10","V4.G10")))
# or use the full list of variable names of each group:
LG_table(ndatas,sets=list(c("V1.G1","V2.G1","V3.G1","V4.G1","V5.G1","V6.G1"),
                           c("V1.G2","V2.G2","V3.G2","V4.G2","V7.G2","V8.G2"),
                           c("V1.G3","V2.G3","V3.G3","V4.G3","V9.G3","V10.G3"),
                           c("V1.G4","V2.G4","V3.G4","V4.G4","V8.G4"),
                           c("V1.G5","V2.G5","V3.G5","V4.G5","V11.G5","V12.G5"),
                           c("V1.G6","V2.G6","V3.G6","V4.G6","V13.G6"),
                           c("V1.G7","V2.G7","V3.G7","V4.G7"),
                           c("V1.G8","V2.G8","V3.G8","V4.G8","V14.G8","V5.G8"),
                           c("V1.G9","V2.G9","V3.G9","V4.G9","V15.G9"),
                           c("V1.G10","V2.G10","V3.G10","V4.G10")))
}

