% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_table.r
\name{LG_table}
\alias{LG_table}
\title{Lg Table}
\usage{
LG_table(dataset, sets)
}
\arguments{
\item{dataset}{a normalized dataframe or matrix}

\item{sets}{list of vector contains vector of indices of each group}
}
\value{
a table of Lg coefficients
}
\description{
Return a table of Lg coefficients between any two subsets of a normalized dataset
}
\examples{
# default 
nadtas <- scale(wine_data)
LG_table(ndatas,sets=list(1:6,7:12,13:18,19:23,24:29,30:34,35:38,39:44,45:49,50:53))
}

