% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadWineData.R
\name{load-wine-info}
\alias{load-wine-info}
\alias{loadWineInfo}
\title{Load Wine Info}
\usage{
loadWineInfo()
}
\value{
Map between V1 --> cattiness etc.
}
\description{
loads wine info (maps of variables V1, etc. to descriptions)
}

