% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{mcmcify}
\alias{mcmcify}
\title{Turn a matrix's columns into informative names}
\usage{
mcmcify(m, name)
}
\arguments{
\item{m}{A fit returned from \code{mfa}}

\item{name}{The name of the parameter}
}
\value{
The input with consistent naming.
}
\description{
Turn a matrix's columns into informative names
}
\keyword{internal}

