% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{mfa}
\alias{mfa}
\title{Multiple factor analysis on a data frame or matrix.}
\usage{
mfa(data, sets, ncomp = NULL, center = TRUE, scale = TRUE)
}
\arguments{
\item{data}{the data set, a data frame or matrix}

\item{sets}{a list of numeric or character vectors indicating the sets of
variables}

\item{center}{either a logical value or a numeric vector of length
equal to the number of active variables in the analysis}

\item{scale}{either a logical value or a numeric vector of length
equal to the number of active variables in the analysis}

\item{ncomps}{integer indicating how many number of components
(i.e. factors) are to be extracted}
}
\value{
an object of class "mfa", containing eigenvalues, factor
    scores, partial factor scores, factor loadings, and other data
    (e.g. weights) for related functions.
}
\description{
Multiple factor analysis on a data frame or matrix.
}

