% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{mfa}
\alias{mfa}
\title{Fit a MFA object}
\usage{
mfa(y, iter = 2000, thin = 1, burn = iter/2, b = 2, pc_initialise = 1,
  collapse = FALSE, seed = 123L, eta_tilde = 5)
}
\value{
Something horrific
}
\description{
Fit a MFA object
}

