% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{mfa}
\alias{mfa}
\title{Multiple Factor Analysis (MFA)}
\usage{
mfa(data, sets, ncomps = NULL, center = TRUE, scale = TRUE)
}
\arguments{
\item{data}{A data frame you'd like to analyze}

\item{sets}{A list of indices}

\item{ncomps}{Number of principle components to retain}

\item{center}{TRUE or FALSE to center data}

\item{scale}{TRUE or FALSE to scale data}
}
\description{
Creates an object of class \code{"mfa"}
}
\examples{
filename = system.file("extdata", "wines.csv", package = "MFA")
d = read.csv(filename, header=TRUE, check.names=FALSE)
s = list(  seq(2,7), seq(8,13), seq(14,19), seq(20,24),
          seq(25,30), seq(31,35), seq(36,39), seq(40,45),
          seq(46,50), seq(51,54) )

a = mfa(d, s)
print(a)

}
\author{
Yulin Chen, Stephanie Wuerth, Eren Bilir, Jim Oreluk
}

