% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{mfa}
\alias{mfa}
\title{mfa}
\usage{
mfa(data, sets, ncomps = NULL, weights = NULL, center = TRUE,
  scale = TRUE, ids = NULL, color = NULL)
}
\arguments{
\item{data}{- a numeric data.frame/matrix containing the data}

\item{sets}{- a list containing vectors of variable names/positions
corresponding to each block}

\item{ncomps}{- number of components to report}

\item{weights}{-}

\item{center}{- logical value: should the data should be centered?}

\item{scale}{- logical value: should the data be scaled?}

\item{ids}{- optional character vector of row names}

\item{color}{- optional character vector for mapping a discrete color scale}
}
\value{
an mfa object contaning the pieces of the resulting analysis
}
\description{
Perform multiple factor analysis
}

