% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{mfa}
\alias{mfa}
\title{mfa}
\usage{
mfa(data, sets, ncomps = 2, weights = NULL, center = TRUE, scale = TRUE,
  ids = NULL, color = NULL)
}
\arguments{
\item{data}{- a numeric data.frame/matrix containing the data}

\item{sets}{- a list containing vectors of variable names/positions
corresponding to each block}

\item{ncomps}{- number of components to report}

\item{weights}{-}

\item{center}{- logical value: should the data should be centered?}

\item{scale}{- logical value: should the data be scaled?}

\item{ids}{- optional character vector of row names}

\item{color}{- optional character vector for mapping a discrete color scale}
}
\value{
an mfa object contaning the pieces of the resulting analysis:
           lambda - the eigenvalues of the compromise
           commonFactorScores - the common factor scores
           partialFactorScores - the partial factor scores
           Q - a matrix of the loadings
           P - the compromise matrix
       In addition, the object with have 8 attributes:
               ncomps - number of components to return while printing 
               ids - the sample ids
               sets - the original list of groups
               colWeights - the vector 'a' of column weights
               rowWeights - the vector 'm' of row weights
               var_names - the variable names
               color - colors for plotting purposes
               boot - a space for optional bootstrapped data
}
\description{
Perform multiple factor analysis
}

