% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFA.R
\name{mfa}
\alias{mfa}
\title{Multiple factor analysis}
\usage{
mfa(data, sets, ncomps = NULL, center = TRUE, scale = TRUE)
}
\arguments{
\item{data}{- data set}

\item{sets}{- list of vectors indicating the sets of variables}

\item{ncomps}{- integer, how many number of components}

\item{center}{- logical argument, whether to center}

\item{scale}{- logical argument, whether to scale}
}
\value{
List containing:

Dimension - Number of the components

Sets - Number of the sets

M - masses for the rows

A - squared singular value derived weights for the k tables

PLin - Vector with weights for lines

PCol - Vector with weights for columns

MatrixA  - Matrix with eigenvalues (Variances)

MatrixU  - Matrix U of the SVD of Matrix Z

MatrixV  - Matriz V of the SVD of Natrix Z

MatrixF  - Compromise Factor Score Matrix

MatrixEFG - Partial Factor Score Matrix

MatrixCCP - Matrix with Correlation of Principal Components with Groups

MatrixEscVar - Matrix of Partial Inertia
}
\description{
This function demostrates basic functionalities of Multiple Factor Analysis
}
\examples{
mfa1 <- mfa(data=mtcars,sets=list(1:3,4:6,7:8,9:11), ncomps = NULL, center = TRUE, scale = TRUE)
}

