% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{mfa}
\alias{mfa}
\title{mfa}
\usage{
mfa(data, sets, ncomps, center = TRUE, scale = TRUE)
}
\arguments{
\item{data}{data set (matrix or data frame)}

\item{sets}{list of vectors indicating the sets of variables}

\item{ncomps}{integer indicating how many components (i.e. factors) should be extracted}

\item{center}{either a logical value or a numeric vector of length equal to # of active vars}

\item{scale}{either a logical value or a numeric vector of length equal to # of active vars}
}
\value{
an object of class coin
}
\description{
Creates an object of class \code{"mfa"}
}
\examples{
# default
mfa1 <- mfa(data, sets=list(2:7,8:13,14:19), ncomps = NULL, center = TRUE, scale = TRUE))
}

