% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_dim.R
\name{obs_dim.mfa}
\alias{obs_dim.mfa}
\title{Contribution: observation to dimension}
\usage{
\method{obs_dim}{mfa}(x)
}
\arguments{
\item{obj}{is an mfa object}
}
\value{
a matrix n-by-m where n is the observations and m is the dimension.
}
\description{
Calculates the constribution of an observation to a dimension.
}

