% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mfa.R
\name{plot compromise}
\alias{plot compromise}
\alias{plot_compromise}
\alias{plot_compromise.mfa}
\title{Plot Compromise}
\usage{
plot_compromise(x, dim1 = 1, dim2 = 2, sz = 2)

\method{plot_compromise}{mfa}(x, dim1 = 1, dim2 = 2, sz = 2)
}
\arguments{
\item{x}{An object of class mfa}

\item{dim1}{dimension for x axis, default 1st component}

\item{dim2}{dimension for y axis, default 2nd component}

\item{sz}{size of point on scatter plot}

\item{x}{An object of class mfa}

\item{dim1}{dimension for x axis, default 1st component}

\item{dim2}{dimension for y axis, default 2nd component}

\item{sz}{size of point on scatter plot}
}
\description{
plots compromise

plots compromise
}
\details{
## Plot Methods
}

