% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compromise.R
\name{plot compromise}
\alias{plot compromise}
\alias{plot_compromise.mfa}
\title{Plot Compromise}
\usage{
\method{plot_compromise}{mfa}(x, dim1 = 1, dim2 = 2, sz = 1,
  obsnames = NULL, textcolor = "black", cexmain = 1, cexlab = 1,
  cexaxis = 0.8, app = FALSE)
}
\arguments{
\item{x}{An object of class mfa}

\item{dim1}{dimension for x axis, default 1st component}

\item{dim2}{dimension for y axis, default 2nd component}

\item{sz}{size of text label in scatter plot}

\item{obsnames}{input labels for your observations if desired}

\item{textcolor}{color of text labeling points, can be one color or a vector of length num obs colors}

\item{cexmain}{size for main title label, default 1, 2 used for app}

\item{cexlab}{size for axis labels, default 1, 1.7 used for app}

\item{cexaxis}{size for axis tick labels, default 0.8}

\item{app}{if true, this is for the shiny app and margins are changed}
}
\description{
plots the compromise, i.e. the common factor scores, for 2 given dimensions of an mfa table
}

