% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ev.R
\name{plot eigenvalue bar graph}
\alias{plot eigenvalue bar graph}
\alias{plot_ev.mfa}
\title{Plot Bar Graph of Eigenvalues}
\usage{
\method{plot_ev}{mfa}(x)
}
\arguments{
\item{x}{An object of class mfa}
}
\description{
plots bar graph of eigenvalues from an mfa object
}

