% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inertia_pie.R
\name{plot inertia pie chart}
\alias{plot inertia pie chart}
\alias{plot_inertia_pie.mfa}
\title{Plot Pie Chart of Inertia}
\usage{
\method{plot_inertia_pie}{mfa}(x)
}
\arguments{
\item{x}{An object of class mfa}
}
\description{
plots a pie chart of percent inertia (from mfa objects eigenvalue table)
}

