% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\docType{methods}
\name{plot,mfa-method}
\alias{plot,mfa-method}
\title{Plot Method for \code{mfa} Object}
\usage{
\S4method{plot}{mfa}(x, dim = c(1, 2), singleoutput = NULL)
}
\arguments{
\item{x}{a R object of \code{mfa}}

\item{dim}{a integer vector of two values}

\item{singleoutput}{a character value among ('eig','com','par') indicating which graph to output}
}
\value{
plots of \code{mfa} Object
}
\description{
Returns plots of eigenvalues, compromise factor scores and partial factor score.
}
\examples{
ndatas<-apply(wine,2,function(x){ (x-mean(x))/norm(x-mean(x),type="2")})
test<-mfa(ndatas,sets=list(1:6,7:12,13:18,19:23,24:29,30:34,35:38,39:44,45:49,50:53),
          center=FALSE,scale=FALSE)

# default
plot(test)

# other dimentions
plot(test,dim=c(2,3))

# single output
plot(test,singleoutput="eig")
plot(test,singleoutput="com")
plot(test,singleoutput="par")
}

