% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_partial_fac.R
\name{plot partial factor scores}
\alias{plot partial factor scores}
\alias{plot_partial_fac.mfa}
\title{Plot Partial Factor Scores}
\usage{
\method{plot_partial_fac}{mfa}(x, table = 1, dim1 = 1, dim2 = 2, sz = 2,
  obsnames = NULL, textcolor = "black")
}
\arguments{
\item{x}{An object of class mfa}

\item{table}{which table to plot (integer 1:K)}

\item{dim1}{dimension for x axis, default 1st component}

\item{dim2}{dimension for y axis, default 2nd component}

\item{sz}{size of text labeling points on scatterplot}

\item{obsnames}{input labels for your observations if desired}

\item{textcolor}{color of text labeling points, either a single color or a vector of colors equal to number of obs in mfa object}
}
\description{
plots partial factor scores for 2 given dimensions/components of a given table in an mfa object
}

