% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loading.R
\name{plot variable loadings}
\alias{plot variable loadings}
\alias{plot_loading.mfa}
\title{Plot Loadings}
\usage{
\method{plot_loading}{mfa}(x, table = 1, dim1 = 1, dim2 = 2, sz = 1,
  varnames = NULL, cexmain = 1, cexlab = 1, cexaxis = 0.8,
  app = FALSE)
}
\arguments{
\item{x}{An object of class mfa}

\item{table}{which table to plot (integer 1:K)}

\item{dim1}{dimension for x axis, default 1st component}

\item{dim2}{dimension for y axis, default 2nd component}

\item{sz}{size of text label in plot}

\item{varnames}{variable labels if available}

\item{cexmain}{size for main title label, default 1, 2 used for app}

\item{cexlab}{size for axis labels, default 1, 1.7 used for app}

\item{cexaxis}{size for axis tick labels, default 0.8}

\item{app}{if TRUE, this is for the shiny app and margins are changed}
}
\description{
plots variable loadings for 2 given dimensions of a given table in an mfa object
}

