% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{plot.mfa}
\alias{plot.mfa}
\title{Plot Method for object of class \code{"mfa"}}
\usage{
\method{plot}{mfa}(mfa, dim1 = 1, dim2 = 2, type = 1, scale_x = 1,
  scale_y = 1, text = NULL, cat = NULL, var = NULL)
}
\arguments{
\item{mfa}{object to be plotted}

\item{dim1}{numeric value that determines the 1st PC the user chooses}

\item{dim2}{numeric value that determines the 2nd PC the user chooses}

\item{type}{numeric value that:
if it is 1, then plot the factor scores
if it is 2, then plot the partial factor scores
if it is 3, then plot the loadings
if it is 4, then plot the partial factor scores and variable loadings}

\item{scale_x}{tunning parameter on the x axis to adjust the scale of the plot}

\item{scale_y}{tunning parameter on the y axis to adjust the scale of the plot}

\item{text}{lables for variables}

\item{cat}{categories to group the observations}

\item{var}{the vector indicating which subtables user wants to see}
}
\value{
Given the choice of type, return graphs of factor scores (the compromise of the tables), of partial factor scores, or of variable loadings
}
\description{
Plots graphs of factor scores (the compromise of the tables), of partial factor scores, or of variable loadings given two dimensions
}
\examples{
cat <- list("New Zealand" = 1:4, "Canada" = 5:8, "France" = 9:12)
dftext <- c("Cat Pee", "Passion Fruit", "Green Pepper",
            "Mineral", "Smoky", "Cirtrus", "Cat Pee",
            "Passion Fruit", "Green Pepper", "Mineral",
            "Tropical", "Leafy", "Cat Pee", "Passion Fruit", "Green Pepper",
            "Mineral", "Grassy", "Flinty", "Cat Pee", "Passion Fruit", "Green Pepper",
            "Mineral", "Leafy", "Cat Pee", "Passion Fruit", "Green Pepper", "Mineral",
            "Vegetal", "Hay", "Cat Pee", "Passion Fruit", "Green Pepper", "Mineral",
            "Melon", "Cat Pee", "Passion Fruit", "Green Pepper", "Mineral", "Cat Pee",
            "Passion Fruit", "Green Pepper", "Mineral", "Grass", "Smoky", "Cat Pee",
            "Passion Fruit", "Green Pepper", "Mineral", "Peach", "Cat Pee",
            "Passion Fruit", "Green Pepper", "Mineral")
plot(mfa = wine, cat = cat, dim1 = 1, dim2 = 2, type = 1, text = dftext)
plot(mfa = wine, cat = cat, dim1 = 1, dim2 = 2, type = 2, text = dftext)
plot(mfa = wine, dim1 = 1, dim2 = 2, type = 3)
}

