% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mfa}
\alias{plot.mfa}
\title{Plot method for 'mfa' object}
\usage{
\method{plot}{mfa}(MFA, labels, dim1, dim2)
}
\arguments{
\item{MFA}{- a 'mfa' object}

\item{labels}{- label for each row}

\item{dim1}{- integer, selected dimension to plot on x-axis}

\item{dim2}{- integer, selected dimension to plot on y-axis}
}
\value{
i) Plot of compromised  factor scores across two selected axes

ii) Plots of partial factor scores & loadings across two selected axes
}
\description{
This function generates i)plot of compromised  factor scores ii)plot of partial factor scores & loadings across two selected axes
}
\examples{
plot(MFA=mfa(mtcars[c(1:2,4:5,8:9,20:21),],sets=list(1:3,4:6,7:8,9:11)),labels=c("Maz1","Maz2","Hor1","Hor2","Merc1","Merc2","Toyota1","Toyota2"),dim1=1,dim2=2)
}

