% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_partial_factor.R
\name{plot_partial_factor}
\alias{plot_partial_factor}
\title{plot_partial_factor}
\usage{
plot_partial_factor(mfa, ...)
}
\arguments{
\item{x}{- the mfa object}

\item{title}{- optional chart title}
}
\value{
a plot with the projection of each observation onto the first two
        extracted components
}
\description{
Plot the compromise scores for the first two extracted dimensions
}

