% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{plot_pfs}
\alias{plot_pfs}
\title{Plot of Partial Factor Scores}
\usage{
plot_pfs(dimension1, dimension2,
  rownames_vec = as.character(1:length(dimension1)), img1 = "black",
  img2 = "black", img3 = "black")
}
\arguments{
\item{dimension1}{vector of class \code{"numeric"}}

\item{dimension2}{vector of class \code{"numeric"}}

\item{rownames_vec}{vector of row labels}

\item{img1}{optional image of class \code{"png"}}

\item{img2}{optional image of class \code{"png"}}

\item{img3}{optional image of class \code{"png"}}
}
\value{
plot of partial factor scores
}
\description{
Takes two dimensions of partial factor scores and returns a plot of the two dimensions
}
\examples{

dimension1 <- rnorm(20)
dimension2 <- rnorm(20)
plot_compromise(dimension1, dimension2)

}

