% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variable_loadings.R
\name{plot_variable_loadings}
\alias{plot_variable_loadings}
\title{plot_variable_loadings}
\usage{
plot_variable_loadings(mfa, ...)
}
\arguments{
\item{x}{- the mfa object}

\item{title}{- chart title}
}
\value{
a plot with the variable loadings for each table
}
\description{
Plot the compromise scores for the first two extracted dimensions
}

