% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{posterior}
\alias{posterior}
\title{Calculate the log-posterior during inference}
\usage{
posterior(y, c, k, pst, tau, gamma, theta, eta, chi, tau_c, r, alpha, beta,
  theta_tilde, eta_tilde, tau_theta, tau_eta, alpha_chi, beta_chi)
}
\description{
Calculate the log-posterior during inference
}

