% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{posterior}
\alias{posterior}
\title{Calculate the log-posterior during inference}
\usage{
posterior(y, c, k, pst, tau, gamma, theta, eta, chi, w, tau_c, r, alpha, beta,
  theta_tilde, eta_tilde, tau_theta, tau_eta, alpha_chi, beta_chi,
  zero_inflation = FALSE, lambda = NULL)
}
\arguments{
\item{y}{Cell-by-gene gene expression matrix}

\item{c}{Factor loading parameter}

\item{k}{Factor loading parameter}

\item{pst}{Pseudotime vector}

\item{tau}{Precision parameter}

\item{gamma}{Branch responsibility parameter}

\item{theta}{Factor loading parameter}

\item{eta}{Factor loading parameter}

\item{chi}{ARD-like precision}

\item{w}{Branch responsibility prior (simplex)}

\item{tau_c}{Hyperparameter}

\item{r}{Hyperparameter}

\item{alpha}{Hyperparamter}

\item{beta}{Hyperparameter}

\item{theta_tilde}{Hyperparameter}

\item{eta_tilde}{Hyperparameter}

\item{tau_theta}{Hyperparameter}

\item{tau_eta}{Hyperparameter}

\item{alpha_chi}{Hyperparameter}

\item{beta_chi}{Hyperparameter}

\item{zero_inflation}{Logical - was zero inflation modelled?}

\item{lambda}{Zero inflation parameter}
}
\value{
The posterior log-likelihood.
}
\description{
Calculate the log-posterior during inference
}

