% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{print.mfa}
\alias{print.mfa}
\title{Print an mfa fit}
\usage{
\method{print}{mfa}(x, ...)
}
\arguments{
\item{x}{An MFA fit returned by \code{mfa}}

\item{...}{Additional arguments}
}
\value{
A string representation of an \code{mfa} object.
}
\description{
Print an mfa fit
}

