% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{print.mfa}
\alias{print.mfa}
\title{Print method for object of class \code{"mfa"}}
\usage{
\method{print}{mfa}(object, subtable = NULL)
}
\arguments{
\item{object}{mfa object}

\item{subtable}{if subtable = NULL then no partial factor score will be printed; if subtable is a numerical value, then the partical factor score for the subtable of this index is printed}
}
\value{
Printed eigenvalues and partical factor score
}
\description{
Print out the eigenvalues and partial factor score of the specified subtable
}
\examples{
# print the eigenvalues
print(wine)

# print the eigenvalues and the partical factor scores for the 6th subtable
print(wine, subtable = 6)
}

