% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFA.R
\name{print.mfa}
\alias{print.mfa}
\title{Print method for 'mfa' object}
\usage{
\method{print}{mfa}(x, ...)
}
\arguments{
\item{x}{- mfa-object}

\item{\dots}{further arguments ignored}
}
\value{
basic information of the 'mfa' object including class, dimension, number of the sets, row mass and table weight of the 'mfa' object
}
\description{
This function defines a print method for "mfa" object
}

