% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{RV}
\alias{RV}
\title{RV Coefficient}
\usage{
RV(table1, table2)
}
\arguments{
\item{table1}{an object of class \code{"matrix"} with same number of rows as table 2}

\item{table2}{an object of class \code{"matrix"} with same number of rows as table 1}
}
\value{
an object of class \code{"numeric"}
}
\description{
Use the RV coefficient to evaluate the similarity between two tables
}
\examples{

table1 <- matrix(rnorm(100), nrow = 10)
table2 <- matrix(rnorm(50), nrow = 10)
RV(table1, table2)

}

