% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{rv}
\alias{rv}
\title{Rv Coefficient - rv}
\usage{
rv(table1, table2)
}
\arguments{
\item{table1}{is the first input table}

\item{table2}{is the second input table}
}
\description{
Calculates Rv coefficient between two tables
}
\examples{
table1 ...
table2 ...
rv(table1,table2) returns the rv coefficient between table1 and table2

}

