% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{rv}
\alias{rv}
\title{Rv Coefficient - rv}
\usage{
rv(table1, table2)
}
\arguments{
\item{x}{An object of class mfa
example rv(table1,table2)}
}
\description{
Calculates Rv coefficient between two tables.
}
\details{
details
}

