% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv.R
\name{rv}
\alias{rv}
\title{Rv Coefficient - rv}
\usage{
rv(table1, table2)
}
\arguments{
\item{table1}{is a list or matrix of the first input table to be included in the analysis}

\item{table2}{is a list or matrix of the second input table to be included in the analysis}
}
\description{
Calculates Rv coefficient between two tables
}
\examples{
filename = system.file("extdata", "wines.csv", package = "MFA")
d = read.csv(filename, header=TRUE, check.names=FALSE)
table1 = d[,seq(2,7)]
table2 = d[,seq(8,13)]
rv(table1,table2)

}

