% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{RV}
\alias{RV}
\title{Rv coefficient of two tables}
\usage{
RV(table1, table2)
}
\arguments{
\item{table1}{data table}

\item{table2}{data table}
}
\value{
the Rv coefficient between table1 and table2
}
\description{
Calculates the Rv coefficients between 2 tables
}
\examples{
t1 <- matrix(c(2, 4, 3, 1, 5, 7))
t2 <- matrix(c(1, 8, 9, 10, 4, 6))
RV(table1 = t1, table2 = t2)
}

