% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV.R
\name{RV}
\alias{RV}
\title{RV}
\usage{
RV(table1, table2)
}
\arguments{
\item{table1}{a data frame or matrix}

\item{table2}{a data frame or matrix}
}
\value{
RV coefficient
}
\description{
To evaluate the similarities between two tables
}
\examples{
t1 = matrix(rnorm(100),nrow=10)
t2 = matrix(rnorm(100),nrow=10)
RV(t1, t2)
}

