% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV.R
\name{RV}
\alias{RV}
\title{RV function}
\usage{
RV(table1, table2)
}
\arguments{
\item{table1}{the first table}

\item{table2}{the second table}
}
\value{
returns the RV coefficient, a measure of similarity for the 2 tables
}
\description{
Calculates RV coefficient for 2 tables, thereby returns a measure of
similarity
}
\examples{
table1 <- matrix(c(1,2,3,4), nrow = 2)
table2 <- matrix(c(9,5,2,8), nrow = 2)
RV(table1,table2)
}
\keyword{RV}
\keyword{measure}
\keyword{similarity}

