% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{RV}
\alias{RV}
\title{RV}
\usage{
RV(table1, table2)
}
\arguments{
\item{table1}{a matrix}

\item{table2}{a matrix which has the same number of rows as table1}
}
\value{
The function will reutrn a number which is a coefficient to evaluate the
        similarity between two table.
}
\description{
RV is the function of MFA package.
             The function aims to analyze the similarity between different table
}

