% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rv.R
\name{RV}
\alias{RV}
\title{Compute RV coefficient between 2 tables}
\usage{
RV(table1, table2, center = TRUE, scale = TRUE)
}
\arguments{
\item{table1}{- matrix of the first table}

\item{table2}{- matrix of the second table}

\item{center}{- logical argument, whether to center}

\item{scale}{- logical argument, whether to scale}
}
\value{
RV - RV coefficients of two tables
}
\description{
This function gets the RV coefficients to study the Between-Table Structure
}
\examples{
mfa1 <- mfa(data=mtcars,sets=list(1:3,4:6,7:8,9:11), ncomps = NULL, center = TRUE, scale = TRUE)
table_1 <- mfa1$MatrixZ[,1:3]
table_2 <- mfa1$MatrixZ[,4:6]
rv.table <- RV(table1=table_1,table2=table_2,center=TRUE,scale=TRUE)
}

