% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV.r
\name{RV}
\alias{RV}
\title{Rv Coefficients}
\usage{
RV(table1, table2)
}
\arguments{
\item{table1}{a normalized data matrix}

\item{table2}{a normalized data matrix}
}
\value{
a value of Rv coefficient between two tables
}
\description{
Return a value of Rv coefficient between two tables
}
\examples{
# default 
x1 <- scale(wine_data[,1:6])
x2 <- scale(wine_data[,7:12])
RV(x1,x2)
}

