% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV.R
\name{RV_table}
\alias{RV_table}
\title{RV coefficient of subsets of a table}
\usage{
RV_table(data, sets)
}
\arguments{
\item{data}{a data frame or matrix}

\item{sets}{a list of numeric or character vectors indicating the sets of
variables}
}
\value{
a symmetric matrix containing the RV coefficients
}

