% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{RV_table}
\alias{RV_table}
\title{RV Coefficient Table}
\usage{
RV_table(dataset, sets)
}
\arguments{
\item{dataset}{an object of class \code{"matrix"} or \code{"data.frame"}}

\item{sets}{object of class \code{"list"}}
}
\value{
an object of class \code{"matrix"}
}
\description{
RV coefficient table contains information of how similar the individual tables are to each other pairwise.
}
\examples{

simulate_data <- matrix(rnorm(200), ncol = 10, byrow = TRUE)
RV_table(simulate_data, sets = list(1:3, 4:5, 6:10))

}

