% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv_table.R
\name{rv_table}
\alias{rv_table}
\title{Rv Coefficient Table - rv_table}
\usage{
rv_table(dataset, sets)
}
\arguments{
\item{dataset}{is the input data in the form of a  data.frame object}

\item{sets}{is a list containing indices (or names) of the variables to be parsed for the analysis}
}
\description{
Calculates Rv coefficient table for multiple sets of a dataset
}
\examples{
filename = system.file("extdata", "wines.csv", package = "MFA")
dataset = read.csv(filename, header=TRUE, check.names=FALSE)
rv_table(dataset, sets = list(1:3, 4:5, 6:10))

}

