% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{rv_table}
\alias{rv_table}
\title{Rv Coefficient Table - rv_table}
\usage{
rv_table(dataset, sets)
}
\arguments{
\item{dataset}{is input data in as a data.frame object.}

\item{sets}{is the list of indices to be parsed from dataset}
}
\description{
Calculates Rv coefficient table multiple sets from a dataset
}
\examples{
dataset = ...
rv_table(dataset, sets = list(1:3, 4:5, 6:10))
returns a 3-by-3 symmetric matrix

}

