% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{rv_table}
\alias{rv_table}
\title{Rv Coefficient Table - rv_table}
\usage{
rv_table(dataset, sets)
}
\arguments{
\item{x}{An object of class mfa
example rv_table(dataset, sets = list(1:3, 4:5, 6:10))
         returns a 3-by-3 symmetric matrix.}
}
\description{
Calculates Rv coefficient table between multiple sets
}
\details{
details
}

