% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{RV_table}
\alias{RV_table}
\title{Rv coefficient of a dataset}
\usage{
RV_table(dataset, sets)
}
\arguments{
\item{dataset}{the grand table from which subtables are extracted}

\item{sets}{list of numeric vectors that extract subtables from the grand table. The ith vector of the list represents the ith subtable extracted}
}
\value{
a matrix with its [i,j] element be the Rv coefficient between the ith subtable and the jth subtable
}
\description{
Calculates the Rv coefficients between the specified subtables in a grand table
}
\examples{
RV_table(datanum, sets = list(sets[[1]], sets[[2]]))
}

