% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV_table.R
\name{RV_table}
\alias{RV_table}
\title{RV_table}
\usage{
RV_table(dataset, sets)
}
\arguments{
\item{dataset}{a data frame or matrix}

\item{sets}{list of vectors indicating the sets of variables (i.e. the blocks).}
}
\value{
a matrix
}
\description{
To evaluate the similarities between each pair of tables among all tables in the dataset
}
\examples{
data(wines)
wines = wines[,2:54]
sets = list(1:6,7:12,13:18,19:23,24:29,30:34,35:38,39:44,45:49,50:53)
RV_table(wines, sets)
}

