% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv.R
\name{RV_table}
\alias{RV_table}
\title{RV_table}
\usage{
RV_table(dataset, setsarg = list(1:2, 3:4))
}
\arguments{
\item{dataset, }{a data-set with numeric variables}

\item{setsarg, }{a list of indices defining subsets of the data-set}
}
\value{
the Rv coefficient matrix between all subsets of data
}
\description{
Computes the Rv coefficient between two tables \code{"rv_table"}
}
\examples{

# With the data-set wines
wines <- read.csv("https://raw.githubusercontent.com/ucb-stat243/stat243-fall-2016/master/problem-sets/final-project/data/wines.csv", stringsAsFactors = FALSE)
sets <- list(2:7, 8:13, 14:19, 20:24, 25:30, 31:35, 36:39, 40:45, 46:50, 51:54)
scaling_vec <- apply(subset(wines, select = unlist(sets)), 2, function(x) sqrt(sum((x - mean(x))^2)))
tables <- data_tables(wines, sets, TRUE, scaling_vec)
sets_shift <- list(1:6, 7:12, 13:18)
dataset1 <- data.frame(cbind(tables[[1]], tables[[2]], tables[[3]]))

RV_dataset <- RV_table(dataset1, sets_shift)



# More simple example
dataset1 <- data.frame(V1 = 1, V2 = 1:10, V3 = c(-1,1), V4 = runif(5), V5 = runif(5))
sets1 <- list(1:2, 3:4)
RV_dataset <- RV_table(dataset1, sets1)



}

