% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV_table.R
\name{RV_table}
\alias{RV_table}
\title{RV Table function}
\usage{
RV_table(mfa, sets = NULL)
}
\arguments{
\item{mfa}{the mfa object}

\item{sets}{list of vectors containing the variables between which the RV
            coefficient
will be calculated}
}
\description{
Function to calculate the RV coefficient between sets of variables
}
\examples{
data(wine)                                    
i <- grep("V", colnames(wine))                
data <- wine[,i]                              
sets <- list(1:6, 7:12, 13:18, 19:23, 24:29,  
            30:34, 35:38, 39:44, 45:49, 50:53)
MFA <- mfa(data, sets)
RV_table(MFA, sets = list(1:6, 19:23, 24:29))
}
\keyword{RV}
\keyword{measure}
\keyword{set}
\keyword{similarity}
\keyword{table}

