% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV_table.R
\name{RV_table}
\alias{RV_table}
\title{RV Table function}
\usage{
RV_table(mfa, sets = NULL)
}
\arguments{
\item{mfa}{the mfa object}

\item{sets}{list of vectors containing the variables between which the RV
            coefficient
will be calculated}
}
\description{
Function to calculate the RV coefficient between sets of variables
}
\examples{
RV_table(data, sets = list(1:3, 4:5, 6:7))
}
\keyword{RV}
\keyword{measure}
\keyword{set}
\keyword{similarity}
\keyword{table}

