% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rv_table.R
\name{RV_table}
\alias{RV_table}
\title{Compute RV coefficient between tables}
\usage{
RV_table(dataset, sets, center = TRUE, scale = TRUE)
}
\arguments{
\item{dataset}{- dataset}

\item{sets}{- the list of table}

\item{center}{- logical argument, whether to center}

\item{scale}{- logical argument, whether to scale}
}
\value{
rv - the matrix of rv coefficient
}
\description{
This function gets the RV coefficients to study the Between-Table Structure
}
\examples{
rv.table <- RV_table(mtcars,sets=list(1:3,4:6,7:8,9:11),center=TRUE,scale=TRUE)
}

