% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV_table.r
\name{RV_table}
\alias{RV_table}
\title{Rv Table}
\usage{
RV_table(dataset, sets)
}
\arguments{
\item{dataset}{a normalized dataframe or matrix}

\item{sets}{list of vector contains vector of indices of each group}
}
\value{
a table of Rv coefficients
}
\description{
Return a table of Rv coefficients between any two subsets of a normalized dataset
}
\examples{
# default 
nadtas <- scale(wine_data)
RV_table(ndatas,sets=list(1:6,7:12,13:18,19:23,24:29,30:34,35:38,39:44,45:49,50:53))
}

