% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{summary.mfa}
\alias{summary.mfa}
\title{Summarise mfa fit}
\usage{
\method{summary}{mfa}(object, ...)
}
\arguments{
\item{object}{An MFA fit returned by a call to \code{mfa}}

\item{...}{Additional arguments}
}
\description{
Summarise mfa fit
}

