% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa_aux.R
\name{summary.mfa}
\alias{summary.mfa}
\title{Prints a table with the singular values (i.e. square root of eigenvalues), the eigenvalues, cumulative,
percentage of intertia, cumulative percentage of inertia, for all the extracted components.}
\usage{
\method{summary}{mfa}(x)
}
\arguments{
\item{x}{object of class mfa.}
}
\value{
None
}
\description{
Prints a table with the singular values (i.e. square root of eigenvalues), the eigenvalues, cumulative,
percentage of intertia, cumulative percentage of inertia, for all the extracted components.
}

