% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_dim.R
\name{table_dim.mfa}
\alias{table_dim.mfa}
\title{Contribution: table to dimension}
\usage{
\method{table_dim}{mfa}(x, ctr_var = 1, numvar_table = 1)
}
\arguments{
\item{x}{is an mfa object}

\item{ctr_var}{is the contribution variable}

\item{numvar_table}{is the variable index in the table}
}
\value{
a matrix n-by-m where n is the tables and m is the dimensions.
}
\description{
Calculates the constribution of a table to the dimension
}

