% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{to_ggmcmc}
\alias{to_ggmcmc}
\title{Turn a trace list to a \code{ggmcmc} object}
\usage{
to_ggmcmc(g)
}
\arguments{
\item{g}{A list of trace matrices}
}
\description{
Turn a trace list to a \code{ggmcmc} object
}

